+++ stories/StylableAutocompleteComposite/AutocompleteCompositeTemplate.js
