+++ stories/StylableAutocompleteComposite/ExampleStandard.js
