+++ stories/StylableAutocompleteComposite/ExampleStandard.scss
