+++ stories/StylableAutocompleteComposite/index.js
